Change executable name to avoid CPU bug
To avoid the high CPU bug observed in #185, the branding in the mozconfig files has been changed to `mercurybrowser` instead of `mercury`. The rest of the branding should remain intact, but the `exe` file will be named `mercurybrowser.exe` instead of `mercury.exe`, which was conflicting with the Gotham Knights application profile with Nvidia.